User commit

+++ src/main/java/com/example/demo/exception/ValidationException.java
@@ -7,7 +7,7 @@ public class ValidationException extends RuntimeException{
 public ValidationException(String message){
 super(message);
 @Override
-       public ValidationEntity getData(Integer id){
+       public ValidationEntity getData(Long id){
                 return validation.findById(id);
                       
                             }
